Implemented Vat services

+++ backend/Server/Models/DTOs/VAT/VatCreateDTO.cs
@@ -3,5 +3,5 @@ namespace backend.Server.Models.DTOs.VAT;
 public class VatCreateDTO
 {
     public required string Name { get; set; }
-    public required decimal Percentage { get; set; }
+    public required float Percentage { get; set; }
 }
